+++ app/models/his/his_ihospital.js
@@ -50,11 +50,8 @@ class HisIHospitalModel {
         else if (wardName) {
             sql.whereLike('ward', `%${wardName}%`);
         }
-        else {
-            sql.where('isactive', 1);
-        }
         return sql
-            .select('code as wardcode', 'ward as wardname', 'standard as std_code', 'bed_nm as bed_normal', 'bed_sp as bed_special', 'ward_type', 'ward_typesub as ward_subtype')
+            .select('code as wardcode', 'ward as wardname', 'standard as std_code', 'bed_nm as bed_normal', 'bed_sp as bed_special', 'ward_type', 'ward_typesub as ward_subtype', 'isactive')
             .limit(maxLimit);
     }
     getDr(db, code, license_no) {