+++ app/models/isonline/his_pmk.model.js
@@ -16,8 +16,16 @@ class HisPmkModel {
                 .where('TABLE_SCHEMA', '=', dbName);
         }
     }
-    testConnect(db) {
-        return db('PATIENTS').select('HN as hn').limit(1);
+    async testConnect(db) {
+        try {
+            const result = await db('PATIENTS').select('HN as hn')
+                .orderBy('HN', 'desc').first();
+            const connection = result && result?.hn ? true : false;
+            return { connection };
+        }
+        catch (error) {
+            throw new Error(error);
+        }
     }
     getPerson(db, columnName, searchText) {
         columnName = columnName === 'hn' ? 'HN' : columnName;