+++ src/models/isonline/his_haos.model.ts
@@ -10,8 +10,15 @@ export class HisHaosModel {
 
     }
 
-    testConnect(db: Knex) {
-        return db('patient').select('hn').limit(1)
+    async testConnect(db: Knex) {
+        try {
+            const result = await db('patient').select('hn').limit(1)
+                .orderBy('hn', 'desc').first();
+            const connection = result && result?.hn ? true : false;
+            return { connection };
+        } catch (error) {
+            throw new Error(error);
+        }
     }
 
     getPerson(db: Knex, columnName: string, searchText: any) {