+++ src/models/isonline/his_ssb.model.ts
@@ -2,6 +2,7 @@ import { Knex } from 'knex';
 import * as moment from 'moment';
 const dbName = process.env.HIS_DB_NAME;
 const maxLimit = 100;
+let hospcode = process.env.HOSPCODE;
 
 export class HisSsbModel {
     getTableName(knex: Knex) {
@@ -10,8 +11,14 @@ export class HisSsbModel {
             .where('TABLE_CATALOG', '=', dbName);
     }
 
-    testConnect(db: Knex) {
-        return db('VW_IS_PERSON').select('hn').limit(1)
+    async testConnect(db: Knex) {
+        try {
+            let result = await db('VW_IS_PERSON').select('hn').first();
+            const connection = result && result?.hn ? true : false;
+            return { connection };
+        } catch (error) {
+            throw new Error(error);
+        }
     }
 
     getPerson(knex: Knex, columnName, searchText) {